debug

+++ climateserv2/processDataRequest.py
@@ -437,7 +437,7 @@ def log_result(retval):
             db.connections.close_all()
             request_progress = Request_Progress.objects.get(request_id=uniqueid)
 
-            update_value = (int(request_progress.progress) - (100/job_length)) - .5
+            update_value = (float(request_progress.progress) - (100/job_length)) - .5
             logger.info('{:.0%} done'.format(update_value))
             # this is so the progress is not set to 100 before the output files are saved to the drive
             # once saved it will update to 100.
